LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/claim-shared-worker-fetch.https-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: SharedWorker


FAIL fetch() in SharedWorker should be intercepted after the client is claimed. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'worker.port')"
